fix backdrop still visible after closing app settings on small screens
